% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_Processing.R
\name{Batch_Processing}
\alias{Batch_Processing}
\title{Process Multiple Mass Spectrometry Data Files}
\usage{
Batch_Processing(
  Spectra_Location = getwd(),
  Sample_Information = F,
  Fragment_Annotation = T,
  Kendricks_Plot = T,
  MS_Spectra = T,
  Find_FA_Attachment = F,
  Find_DB_from_FA = F,
  Loss_Annotation = T,
  Find_DBs = T,
  Find_Additional_DB = T,
  Create_Summary = T,
  File.Column = "File.Name",
  Analyte.Column = "Letter_Code",
  Precursor.Column = "Precursor",
  Loss.Column = "Add_Losses",
  Composition.Column = "Analyte"
)
}
\arguments{
\item{Spectra_Location}{A character string specifying the location of the spectra files. Defaults to the current working directory.}

\item{Sample_Information}{A character string or logical indicating whether sample information is provided (e.g., an Excel file). Defaults to FALSE.}

\item{Fragment_Annotation}{A logical indicating whether to perform fragment annotation. Defaults to TRUE.}

\item{Kendricks_Plot}{A logical indicating whether to generate Kendrick plots. Defaults to TRUE.}

\item{MS_Spectra}{A logical indicating whether to generate mass spectrum plots. Defaults to TRUE.}

\item{Find_FA_Attachment}{A logical indicating whether to find and extract fatty acid peaks. Defaults to FALSE.}

\item{Find_DB_from_FA}{A logical indicating whether to find double bonds from fatty acid data. Defaults to FALSE.}

\item{Loss_Annotation}{A logical indicating whether to annotate elemental formula losses. Defaults to TRUE.}

\item{Find_DBs}{A logical indicating whether to find double bonds based on fragmentation data. Defaults to TRUE.}

\item{Find_Additional_DB}{A logical indicating whether to find additional double bonds based on losses. Defaults to TRUE.}

\item{Create_Summary}{A logical indicating whether to create a summary of the results. Defaults to TRUE.}

\item{File.Column}{A character string specifying the column name in the sample information that contains file names. Defaults to "File.Name".}

\item{Analyte.Column}{A character string specifying the column name for analyte codes. Defaults to "Letter_Code".}

\item{Precursor.Column}{A character string specifying the column name for precursor masses. Defaults to "Precursor".}

\item{Loss.Column}{A character string specifying the column name for additional losses. Defaults to "Add_Losses".}

\item{Composition.Column}{A character string specifying the column name for analyte compositions. Defaults to "Analyte".}
}
\value{
A list containing processed data and plots for each spectrum file. The list includes:
\itemize{
\item Input_Data: The loaded data for the current spectrum.
\item Precursor_Data: The precursor data.
\item Annotated_Data: The annotated fragment data.
\item Kendricks_Data: The Kendrick transformed data.
\item MG_Fragment_Data: Data related to fatty acid fragments.
\item Annotated_Loss_Data: Data related to annotated losses.
\item DG_Fragment_Data: Data for double bond fragments.
\item DB_Data: Data for double bonds.
\item DB_After_Loss_Data: Data for additional double bonds after loss.
}
}
\description{
This function processes multiple mass spectrometry data files based on user-defined parameters.
}
\examples{
# Basic usage
results <- Batch_Processing(Spectra_Location = "path/to/spectra", Sample_Information = "path/to/sample_info.xlsx")

}
