% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{DB_gg_func}
\alias{DB_gg_func}
\title{Generate a Heatmap of Double Bond Data}
\usage{
DB_gg_func(
  data,
  Oxygen_Range = c(0:6),
  filling = "scaled",
  H_Count_min = -10,
  H_Count_Max = 10,
  Carbon_Pos_min = 2,
  Carbon_Pos_Max = 20
)
}
\arguments{
\item{data}{A data frame containing the information to be plotted. It must include
the columns \code{delta_H}, \code{C} (Carbon_Position), \code{Height}, and \code{O} (oxygen levels).}

\item{Oxygen_Range}{A vector specifying the range of oxygen values to include in
the plot. Default is 0 to 6.}

\item{filling}{A string determining how the tiles in the heatmap are filled. Options
are "scaled" for scaled intensity or "log10" of Height. Defaults to "scaled".}

\item{H_Count_min}{Minimum value for \code{delta_H} to include in the plot. Default is -10.}

\item{H_Count_Max}{Maximum value for \code{delta_H} to include in the plot. Default is 10.}

\item{Carbon_Pos_min}{Minimum value for \code{C} (Carbon_Position) to include in the plot.
Default is 2.}

\item{Carbon_Pos_Max}{Maximum value for \code{C} (Carbon_Position) to include in the plot.
Default is 20.}
}
\value{
A heatmap generated by ggplot2, showing the relationship between \code{delta_H},
\code{Carbon_Position}, and intensity values based on the specified filling method.
}
\description{
This function creates a heatmap visualizing double bond data based on specified
parameters. It can fill tiles based on either scaled intensity or the logarithm
of height values.
}
\examples{
# Assuming 'data_frame' is a data frame with the necessary columns
heatmap_plot <- DB_gg_func(data = data_frame, Oxygen_Range = 0:4, filling = "log10")
print(heatmap_plot)

}
