% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Extract_DB_Plots}
\alias{Extract_DB_Plots}
\title{Extract Double Bond Plot Objects}
\usage{
Extract_DB_Plots(Data_List, Figure_Pos = 3, delta_H = "even")
}
\arguments{
\item{Data_List}{A list of plot objects, potentially containing entries labeled
as "No Peak". This list is generated by the \code{Find_Double_Bonds} function.}

\item{Figure_Pos}{An integer indicating the position of the plot to extract from the
list. Defaults to 3.}

\item{delta_H}{A specification for filtering plots based on the number of double bonds.
It can be:
\itemize{
\item A numeric value to extract specific double bond counts.
\item A string "even" to filter for even counts.
\item A string "odd" to filter for odd counts. Defaults to "even".
}}
}
\value{
A wrapped plot object containing the extracted double bond plots. If no valid
data entries remain after filtering, a message indicating "No Data" will be returned.
}
\description{
This function extracts double bond plot objects created by the \code{Find_Double_Bonds}
functions. It filters and returns a series of plots based on specified criteria,
wrapped into a single output.
}
\examples{
# Assuming 'data_list' is a list of plot objects obtained from Find_Double_Bonds
extracted_plots <- Extract_DB_Plots(Data_List = data_list, Figure_Pos = 3, delta_H = "even")
print(extracted_plots)

}
