% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Extract_FAs}
\alias{Extract_FAs}
\title{Extract Carbon Numbers of Fatty Acid Fragments}
\usage{
Extract_FAs(
  Data,
  FA_Peaks,
  Fa_Number = (Precursor_DF$O - 3),
  Precursor_Df,
  Backbone_C = 3,
  O_Count = c(1:3)
)
}
\arguments{
\item{Data}{A data frame containing the data to be filtered, including columns for
\code{C} (number of carbon atoms), \code{O} (number of oxygen atoms), and other relevant
measurements.}

\item{FA_Peaks}{A numeric vector of fatty acid peaks identified in the dataset.}

\item{Fa_Number}{Numeric. The number of fatty acids to consider in combinations
(default: calculated as \code{Precursor_DF$O - 3}).}

\item{Precursor_Df}{A data frame containing precursor information, including carbon counts.}

\item{Backbone_C}{Numeric. The number of carbons in the glycerol backbone (default: 3).}

\item{O_Count}{A numeric vector of oxygen counts to filter the data (default: 1:3).}
}
\value{
A data frame containing entries corresponding to identified potential fatty
acids, including columns for \code{FA} (indicating potential fatty acid) and \code{delta_H}.
Returns NA if no peaks are found.
}
\description{
This function extracts the carbon number of fatty acid fragments based on
monoacylglycerol (MG) and fatty acid fragments. It matches combinations of carbon
atoms to the precursor carbon count after glycerol carbon deduction.
}
\examples{
# Example usage of Extract_FAs function
extracted_fas <- Extract_FAs(Data = mg_data, FA_Peaks = fa_peaks)

}
