% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Extract_List}
\alias{Extract_List}
\title{Extract the n-th Element from Each Item in a List}
\usage{
Extract_List(lst, n)
}
\arguments{
\item{lst}{A list from which elements will be extracted. Each item in the list
should be a vector or a list itself.}

\item{n}{An integer representing the index of the element to extract from each
item in the list. Indices start from 1.}
}
\value{
A vector containing the n-th elements extracted from each item in the
list. If n exceeds the length of any item, NA will be returned for that item.
}
\description{
This function extracts the n-th element from each item in a given list.
}
\examples{
# Example usage of Extract_List function
my_list <- list(c(1, 2, 3), c(4, 5, 6), c(7, 8, 9))
result <- Extract_List(my_list, 2)
# result will be c(2, 5, 8)

}
