% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Find_DB}
\alias{Find_DB}
\title{Identify and Plot Data Related to Double Bonds}
\usage{
Find_DB(
  Data,
  delta_H = c(0, 2, 4, 6),
  Threeshold = 10,
  Oxygen = 0,
  mincut = 0.382,
  xstart = 0,
  minpeak = 0.1,
  text_size = 10
)
}
\arguments{
\item{Data}{A data frame containing columns for carbon counts (C), height,
slope, and delta_H values. The data should include measurements related to
fatty acids.}

\item{delta_H}{A numeric vector of delta_H values to filter the data.
Default is c(0, 2, 4, 6).}

\item{Threeshold}{A numeric value representing the minimum height threshold
for filtering data. Default is 10.}

\item{Oxygen}{A numeric value representing the oxygen count to filter data.
Default is 0.}

\item{mincut}{A numeric value used as a threshold for peak detection.
Default is 0.382.}

\item{xstart}{A numeric value representing the starting point for peak
detection. Default is 0.}

\item{minpeak}{A numeric value for the minimum height of detected peaks.
Default is 0.1.}

\item{text_size}{A numeric value for text size in plots. Default is 10.}
}
\value{
A list containing:
\itemize{
\item \code{Peaks_Height_Based}: Data related to height peaks.
\item \code{Peaks_Slope_Based}: Data related to slope peaks.
\item \code{DB_Figure}: A combined plot of height and slope for each delta_H value.
}
}
\description{
This function identifies and plots data related to double bonds in a given
dataset. It filters the data based on specified conditions and creates slope
and height plots.
}
\examples{
# Example usage of Find_DB function
result <- Find_DB(Data = your_data_frame)
# This will filter your_data_frame and create plots for identified double bonds.

}
