% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Find_DB2}
\alias{Find_DB2}
\title{Add Secondary Double Bond Series Based on Specified Losses}
\usage{
Find_DB2(
  Data = Data,
  Losses = "",
  DBE_min = -20,
  C0 = TRUE,
  FA_Peak_vec = FA_Loss_Peaks
)
}
\arguments{
\item{Data}{A data frame containing the original data. Default is a variable
named \code{Data}.}

\item{Losses}{A string specifying the losses (chemical formulas or mass values)
to be processed. Default is an empty string, which calculates losses based
on FA peaks.}

\item{DBE_min}{Minimum double bond equivalent to consider. Default is -20.}

\item{C0}{A boolean indicating whether to keep carbon count of zero. Default
is TRUE.}

\item{FA_Peak_vec}{A vector of fatty acid peaks. Default is \code{FA_Loss_Peaks}.}
}
\value{
A list containing processed data frames and any invalid inputs. The
list may contain:
\itemize{
\item \code{Sec_Data_DB_filtered}: Filtered data for calculated losses.
\item \code{Sec_Data_Loss_H3}: Filtered data with carbon count equal to zero.
\item Messages for invalid inputs if applicable.
}
}
\description{
This function adds secondary double bond series to the provided data based
on specified losses. It calculates double bond equivalents and filters the
data accordingly.
}
\examples{
# Example usage of Find_DB2 function
result <- Find_DB2(Data = my_data, Losses = "C18H36O2")

}
