% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Kendricks_Transformation}
\alias{Kendricks_Transformation}
\title{Kendrick Mass Defect Transformation}
\usage{
Kendricks_Transformation(
  Kendricks_Data,
  Kendrick_Defect = 14/14.01565,
  Mass.Charge = Mass.Charge
)
}
\arguments{
\item{Kendricks_Data}{A data frame containing spectra data with at least
columns for mass/charge values and intensity.}

\item{Kendrick_Defect}{Numeric. Kendrick mass defect, usually CH2 group (default: 14 / 14.01565).}

\item{Mass.Charge}{Character. Column name for mass/charge values in \code{Kendricks_Data}.}
}
\value{
A data frame with the transformed Kendrick mass defect values,
relative intensities, and other calculated columns for further analysis.
}
\description{
Performs the Kendrick mass defect transformation on the input spectra data
and calculates the relative intensities of fragment ions.
}
\examples{
# Example of transforming data with default Kendrick defect for CH2 group
transformed_data <- Kendricks_Transformation(Kendricks_Data = your_data)

}
