% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_Examples.R
\name{Load_Example_Path}
\alias{Load_Example_Path}
\title{Load Example Spectra Path}
\usage{
Load_Example_Path(Sample, file.type = ".txt")
}
\arguments{
\item{Sample}{A character string specifying the sample name. This is used
to construct the filename for the example spectrum.}

\item{file.type}{A character string specifying the file type. Default is ".txt".}
}
\value{
A character string representing the path to the specified example
spectrum file.
}
\description{
This function generates the path to a specified example spectrum file
from the "Example Files" directory of the MsRadaR package based on
the provided sample name and file type.
}
\examples{
# Load the path for the  sample "LLL.txt"
Example_Spectra_Path <- Load_Example_Path(Sample = "LLL")

}
