% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Load_Parameters}
\alias{Load_Parameters}
\title{Load Default Parameters for Elemental Formula Calculation}
\usage{
Load_Parameters(
  Threshold = 2,
  mDaError = 20,
  Kendricks_Defect = 14/14.01565,
  minFormula_Lipid = "C3H8O2",
  min_formula = "C0H1",
  max_Formula = "C200H800O6",
  max_Formula_FA = "C20H40O4",
  Allowed_Elements = "CHO"
)
}
\arguments{
\item{Threshold}{Numeric. Intensity threshold for data loading (default: 2).}

\item{mDaError}{Numeric. Allowed mass error in mDa (default: 20).}

\item{Kendricks_Defect}{Numeric. Kendrick mass defect (default: 14 / 14.01565).}

\item{minFormula_Lipid}{Character. Minimum elemental formula for lipids (default: "C3H8O2").}

\item{min_formula}{Character. Minimum elemental formula (default: "C0H1").}

\item{max_Formula}{Character. Maximum elemental formula (default: "C200H800O6").}

\item{max_Formula_FA}{Character. Maximum elemental formula for fatty acids (default: "C20H40O4").}

\item{Allowed_Elements}{Character. String of allowed elements (default: "CHO").}
}
\value{
Sets global variables for threshold, allowed mass error, Kendrick defect,
and elemental formula constraints based on specified inputs.
}
\description{
Loads default parameters for elemental formula calculation and data loading.
These settings can be adjusted according to user-defined preferences.
}
\examples{
Load_Parameters()
Load_Parameters(Threshold = 5, mDaError = 15)

}
