% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{Name_to_Formula}
\alias{Name_to_Formula}
\title{Convert Fatty Acid Name to Elemental Formula}
\usage{
Name_to_Formula(Name)
}
\arguments{
\item{Name}{A character string representing the fatty acid names in the format
"number of carbons:number of double bonds" (e.g., "18:1").}
}
\value{
A character vector containing the elemental formula in the format
"C\link{number of carbons}H\link{number of hydrogens}O2".
}
\description{
This function takes a string representing fatty acid names (e.g., "18:1") and
converts it into its corresponding elemental formula.
}
\examples{
# Example usage of Name_to_Formula function
formula <- Name_to_Formula("18:1")

}
