% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Processing_Functions.R
\name{peakwindow_adapt}
\alias{peakwindow_adapt}
\title{Identify Maxima within Series with Low Data Points}
\usage{
peakwindow_adapt(
  x,
  y = NULL,
  xstart = 0,
  xmax = max(x),
  minpeak = 0.1,
  mincut = 0.382
)
}
\arguments{
\item{x}{Numeric vector. The x-coordinates of the data points (e.g., time or
m/z values).}

\item{y}{Numeric vector. The y-coordinates of the data points (e.g., intensity
values).}

\item{xstart}{Numeric. The starting x-coordinate to consider for peak
detection (default: 0).}

\item{xmax}{Numeric. The maximum x-coordinate for peak detection (default: max(x)).}

\item{minpeak}{Numeric. The minimum threshold for peak detection, specified as
a fraction of the maximum y value (default: 0.1).}

\item{mincut}{Numeric. Threshold for cutting insignificant peaks; values below
this fraction of the minimum value in each peak range are ignored (default: 0.382).}
}
\value{
A list of class \code{cardiPeakwindow}, containing the following components:
\item{peaks}{A data frame of detected peaks with indices, left and right bounds,
x and y coordinates, and intensities.}
\item{data}{A data frame containing the original x and y data points.}
\item{smd.max.index}{The index of the peak with the maximum smoothed y value.}
\item{smd.max.x}{The x-coordinate of the peak with the maximum smoothed y value.}
\item{smd.indices}{Indices of the smoothed region.}
\item{smd.x}{x-coordinates of the smoothed region.}
\item{smd.y}{y-coordinates of the smoothed region.}
\item{peakid}{Peak identifiers for each point in the y data.}
}
\description{
This function is adapted from the \code{peakwindow} function in the \code{cardiates}
package to identify peaks within series that contain low data points. It
allows for the detection of maximum points with flexibility in peak limits.
}
\examples{
# Example usage of peakwindow_adapt function
peaks <- peakwindow_adapt(x = seq(1, 100), y = rnorm(100))

}
