% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Plot_DB_from_FA}
\alias{Plot_DB_from_FA}
\title{Generate a Plot of Carbon Counts vs Heights for Elemental Formulas}
\usage{
Plot_DB_from_FA(
  Data = Elemental_Formula_df,
  column = "L_R",
  Lower_H_Limit = -2,
  Upper_H_Limit = 20,
  Thres = 5,
  Max_C = max(FA_DF$C),
  Oxygen_Range = 0:3,
  point_size = 1,
  alpha_v = 0.35
)
}
\arguments{
\item{Data}{A data frame (default is \code{Elemental_Formula_df}) containing the elemental formula data.}

\item{column}{A string indicating which column in the data represents the delta H values.
Default is "L_R".}

\item{Lower_H_Limit}{Minimum threshold for \code{H_Delta} to include in the plot. Default is -2.}

\item{Upper_H_Limit}{Maximum threshold for \code{H_Delta} to include in the plot. Default is 20.}

\item{Thres}{Minimum threshold for \code{Height} to include in the plot. Default is 5.}

\item{Max_C}{Maximum carbon count to include in the plot. Default is the maximum value in \code{FA_Peaks}.}

\item{Oxygen_Range}{A vector indicating the range of oxygen values to include. Default is 0 to 3.}

\item{point_size}{Size of the points in the plot. Default is 1.}

\item{alpha_v}{Transparency level for the lines in the plot. Default is 0.35.}
}
\value{
A ggplot object visualizing the relationship between carbon counts and heights
of elemental formulas based on the specified filtering criteria. If no data meets the
filtering criteria, a message is returned indicating this.
}
\description{
This function creates a plot visualizing the relationship between carbon counts (C)
and heights of elemental formulas, conditioned on specific delta H values and other
parameters. It allows for filtering based on thresholds for height, delta H, and oxygen
levels.
}
\examples{
# Assuming 'Elemental_Formula_df' is available and contains necessary columns
plot <- Plot_DB_from_FA(Data = Elemental_Formula_df, Lower_H_Limit = 0)
print(plot)

}
