% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Plot_FAs}
\alias{Plot_FAs}
\title{Generate a Bar Plot for Monoacylglycerol and Fatty Acid Fragments}
\usage{
Plot_FAs(Data, xlim1 = -10, xlim2 = 5, Oxygen_Range = 0:10)
}
\arguments{
\item{Data}{A data frame containing at least the following columns:
\itemize{
\item \code{H}: Hydrogen count.
\item \code{delta_H}: Delta hydrogen count representing changes due to losses.
\item \code{Height}: Height values for the data points.
\item \code{O}: Oxygen count.
\item \code{C}: Carbon count.
\item \code{Mass.Charge}: Mass charge information to be displayed on the plot.
}}

\item{xlim1}{A numeric value representing the lower limit for \code{delta_H}. Defaults to -10.}

\item{xlim2}{A numeric value representing the upper limit for \code{delta_H}. Defaults to 5.}

\item{Oxygen_Range}{A numeric vector specifying the range of oxygen counts to include.
Defaults to 0:10 (i.e., includes oxygen counts from 0 to 10).}
}
\value{
A ggplot object representing the bar plot of monoacylglycerol and fatty acid fragments.
If no data matches the criteria, a message indicating this is returned.
}
\description{
This function creates a bar plot to visualize monoacylglycerol and fatty acid
fragments based on their delta hydrogen counts (\code{delta_H}) and height values,
categorized by carbon and oxygen counts. It distinguishes between different
fragmentation mechanisms (radical/neutral losses) using even/odd hydrogen counts.
}
\examples{
# Assuming 'fa_data' is a data frame containing the required columns
plot <- Plot_FAs(Data = fa_data)
print(plot)

}
