% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Plot_Kendricks_Plot}
\alias{Plot_Kendricks_Plot}
\title{Generate a Kendrick Plot}
\usage{
Plot_Kendricks_Plot(
  data,
  toggle_alpha = 0.05,
  Rel.Thres = 1/100,
  Abs.Thres = Signal_Noise,
  plot_legend = TRUE,
  point_size = 1
)
}
\arguments{
\item{data}{A data frame containing at least the following columns:
\itemize{
\item \code{Height}: Signal intensity for the data points.
\item \code{Rel.Int}: Relative intensity of the data points.
\item \code{O}: Oxygen content.
\item \code{Ke_CH2}: Kendrick Mass values.
\item \code{KDM_CH2}: Kendrick Defect values.
}}

\item{toggle_alpha}{A numeric value indicating whether to adjust alpha levels for points
below the relative intensity threshold. Defaults to 0.05.}

\item{Rel.Thres}{A numeric value representing the relative intensity threshold for
adjusting alpha. Defaults to 0.01.}

\item{Abs.Thres}{A numeric value representing the absolute signal threshold for data points.}

\item{plot_legend}{A logical value indicating whether to display the legend on the plot.
Defaults to TRUE.}

\item{point_size}{A numeric value for the size of the points in the plot. Defaults to 1.}
}
\value{
A ggplot object representing the Kendrick plot. The plot visualizes Kendrick Mass
versus Kendrick Defect, with points colored by oxygen content and alpha transparency
based on relative intensity.
}
\description{
This function creates a Kendrick plot using ggplot2, visualizing the relationship
between Kendrick Mass (Kendrick Defect) based on CH2. The plot differentiates data points
by color according to their oxygen content and adjusts point transparency based on
relative intensity thresholds.
}
\examples{
# Assuming 'kendrick_data' is a data frame containing the required columns
plot <- Plot_Kendricks_Plot(data = kendrick_data)
print(plot)

}
