% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting_Functions.R
\name{Plot_MS2}
\alias{Plot_MS2}
\title{Generate an Annotated MS/MS Plot}
\usage{
Plot_MS2(
  Data,
  Graph_Splitting_Ratio = 1.1,
  Axis_Split_Size = 5,
  Split_Text_Size = 1,
  Label_Size = 8,
  Line_Size = 1,
  Rel_Threeshold = 1
)
}
\arguments{
\item{Data}{A data frame containing at least the following columns:
\itemize{
\item \code{Rel.Int}: Relative intensity values of the fragments.
\item \code{Height}: Height values corresponding to the fragments (for reference).
\item \code{Mass.Charge}: Mass-to-charge ratio (m/z) for each fragment.
\item \code{O}: Oxygen count for each fragment.
}}

\item{Graph_Splitting_Ratio}{A numeric value that defines the ratio used to
determine whether to split the graph for lower intensity fragments. Defaults to 1.1.}

\item{Axis_Split_Size}{A numeric value controlling the size of the axis split line.
Defaults to 5.}

\item{Split_Text_Size}{A numeric value that sets the size of the split label text.
Defaults to 1.}

\item{Label_Size}{A numeric value that specifies the size of the labels on the plot.
Defaults to 8.}

\item{Line_Size}{A numeric value that determines the size of the line segments in the plot.
Defaults to 1.}

\item{Rel_Threeshold}{A numeric threshold for relative intensity filtering.
Defaults to 1.}
}
\value{
A ggplot object representing the annotated MS/MS plot. If significant intensity
differences are not present, an unzoomed plot will be returned.
}
\description{
This function creates an annotated MS/MS plot, visualizing relative intensities
of fragments based on mass-to-charge ratio (m/z) and oxygen count. The plot
automatically adjusts the y-axis to zoom in on lower intensity fragments when
there is a significant difference in intensities.
}
\examples{
# Assuming 'ms_data' is a data frame containing the required columns
plot <- Plot_MS2(Data = ms_data)
print(plot)

}
